+++ src/v1/BasicTypes/Numbers/validateNumberRange.ts
@@ -31,8 +31,38 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
+import { NumberOutOfRangeError } from "../../Errors";
+import { AppErrorOr } from "../../OptionTypes";
+import { DataPath } from "../../SupportingTypes";
 
-export * from "./isNumber";
-export * from "./mustBeNumber";
-export * from "./validateNumber";
-export * from "./validateNumberRange";
+/**
+ * `validateNumberRange()` is a {@link TypeValidator}. Use it to prove that
+ * the given input falls within the range of values you require.
+ *
+ * @param path
+ * @param input
+ *
+ * @category BasicTypes
+ */
+export function validateNumberRange(
+    path: DataPath,
+    input: number,
+    minInc: number,
+    maxInc: number,
+    { rangeError = NumberOutOfRangeError } = {}
+): AppErrorOr<number> {
+    // just your basic range check
+    if (input < minInc || input > maxInc) {
+        return new rangeError({
+            public: {
+                dataPath: path,
+                input,
+                minInc,
+                maxInc,
+            }
+        });
+    }
+
+    // all done
+    return input;
+}